api/pyproject.toml
[project]
name = "apollonia-api"
version = "0.1.0"
description = "RESTful API and GraphQL service for Apollonia"
requires-python = ">=3.12,<3.13"

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.25.0",
    "strawberry-graphql[fastapi]>=0.217.0",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "redis>=5.0.0",
    "aiofiles>=24.1.0",
    "httpx>=0.25.0",
    "orjson>=3.9.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "prometheus-client>=0.19.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
